Separated actions from action creators
