+++ src/store/actions/actions.js
@@ -1,6 +1,4 @@
-import axios from 'axios';
-import { axiosWithAuth } from '../../axiosAuth';
-
+// ACTIONS
 export const LOGIN_START = 'LOGIN_START';
 export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
 export const LOGIN_FAILURE = 'LOGIN_FAILURE';
@@ -25,135 +23,10 @@ export const UPDATE_SUCCESS = 'UPDATE_SUCCESS';
 export const ADD_CATEGORY = 'ADD_CATEGORY';
 export const DELETE_CATEGORY = 'DELETE_CATEGORY';
 export const REMOVE_CATEGORY = 'REMOVE_CATEGORY';
-export const ADD_CATEGORY_LINK = 'ADD_CATEGORY_LINK';
 
 export const SEARCH_TAB = 'SEARCH_TAB';
 export const UNDO_SEARCH = 'UNDO_SEARCH';
 
 export const TAB_VISITED = 'TAB_VISITED';
 
-//ACTIONS CREATOR
 
-const loginAPI = "https://tabless-thursday-backend.herokuapp.com/api/login";
-
-export const login = ({ username, password, email }) => dispatch => {
-    const credentials = { username, password, email };
-
-    dispatch({ type: LOGIN_START });
-
-    return axios
-                .post(loginAPI, credentials)
-                .then(response => {
-                    localStorage.setItem('token', response.data.token);
-                    localStorage.setItem('userID', response.data.user.id);
-                    localStorage.setItem('userLogged', true);
-                    dispatch({ type: LOGIN_SUCCESS, payload: response.data.user }); 
-                })
-                .catch(error => {
-                    debugger
-                })
-                .finally(() => {
-                    dispatch({ type: LOGIN_END });
-                })
-}
-
-export const logOutUser = () => dispatch => {
-    localStorage.clear();
-    dispatch({ type: LOGOUT_USER });
-}
-
-const registerAPI = "https://tabless-thursday-backend.herokuapp.com/api/register/";
-
-
-export const registerUser = ({ username, password, email }) => dispatch => {
-    // const credentials = { username, password, email }
-
-    return  axios
-                .post(registerAPI, { username, password, email })
-                .then(response => {
-                    localStorage.setItem('user', JSON.stringify(response.data.saved));
-                    dispatch({ type: REGISTER_SUCCESS, payload: response.data.saved })
-                })
-                .catch(error => {
-                    debugger
-                })
-}
-
-export const getUserTabs = () => dispatch => {
-    let id = localStorage.getItem('userID');
-
-    dispatch({ type: FETCH_START });
-
-    axiosWithAuth()
-        .get(`https://tabless-thursday-backend.herokuapp.com/api/users/${id}`)
-        .then(response => {
-            dispatch({ type: FETCH_SUCCESS, payload: response.data.tabs })
-        })
-        .catch(error => {
-            debugger
-        })
-        .finally(() => {
-            dispatch({ type: FETCH_END })
-        })
-}
-
-const postTabAPI = "https://tabless-thursday-backend.herokuapp.com/api/tabs";
-
-export const postUserTab = (tabInfo) => dispatch => {
-    axiosWithAuth()
-        .post(postTabAPI, tabInfo) 
-        .then(response => {
-            dispatch({ type: CREATE_TAB, payload: response.data[0] })
-            // getUserTabs();
-        })
-        .catch(error => {
-            debugger
-        })
-}
-
-export const deleteTab = (id) => dispatch => {
-    axiosWithAuth()
-        .delete(`https://tabless-thursday-backend.herokuapp.com/api/tabs/${id}`)
-        .then(response => {
-            dispatch({ type: DELETE_TAB, payload: response.data.message, tabId: id });
-        })
-        .catch(error => {
-            debugger
-        })
-}
-
-export const updateTab = (id, tabInfo) => dispatch => {
-
-    axiosWithAuth()
-        .put(`https://tabless-thursday-backend.herokuapp.com/api/tabs/${id}`, tabInfo)
-        .then(response => {
-            dispatch({ type: UPDATE_SUCCESS, payload: JSON.parse(response.config.data), tabId: id });
-        })
-        .catch(error => {
-            debugger
-        })
-}
-
-export const addCategory = (category) => dispatch => {
-    dispatch({ type: ADD_CATEGORY, payload: category });
-}
- 
-export const addCategoryLink = (nameLink) => dispatch => {
-    dispatch({ type: ADD_CATEGORY, payload: nameLink });
-}
-
-export const removeCategory = (category) => dispatch => {
-    dispatch({ type: REMOVE_CATEGORY, payload: category });
-}
-
-export const searchTab = (searchResult) => dispatch => {
-    dispatch({ type: SEARCH_TAB, payload: searchResult });
-}
-
-export const undoSearch = () => dispatch => {
-    dispatch({ type: UNDO_SEARCH });
-}
-
-export const tabVisited = (tabId) => dispatch => {
-    dispatch({ type: TAB_VISITED, payload: tabId});
-}
